+++ Assets/Scripts/Commons.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1687c43fa5dc372429cdf88315465e19
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
